+++ src/pages/Login/Login.styled.js
@@ -2,8 +2,6 @@ import styled, { css } from 'styled-components';
 import pxToRem from '@utils/pxToRem';
 
 import eyeIcon from '@assets/svgs/icons/eyelnvisible.svg?react';
-import googleLogo from '@assets/svgs/icons/google_logo.svg?react';
-import kakaoLogo from '@assets/svgs/icons/kakao_logo.svg?react';
 
 export const LoginContainer = styled.div`
   width: ${pxToRem(380)};
@@ -93,57 +91,6 @@ export const SocialDivider = styled.div`
 
   border-radius: 4px;
   background-color: ${({ theme }) => theme.colors.MG_Grayscale.Gray_2};
-`;
-
-export const SocialLoginContainer = styled.div`
-  width: 100%;
-  height: fit-content;
-
-  display: flex;
-  flex-direction: column;
-  gap: ${pxToRem(16)};
-`;
-
-const socialLoginBtnStyle = css`
-  width: 100%;
-  height: fit-content;
-  padding: ${pxToRem(16)} ${pxToRem(24)};
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: ${pxToRem(24)};
-
-  border-radius: 8px;
-`;
-
-export const GoogleLoginBtn = styled.button`
-  ${socialLoginBtnStyle}
-  background-color: ${({ theme }) => theme.colors.MG_Grayscale.White};
-`;
-
-export const KakaoLoginBtn = styled.button`
-  ${socialLoginBtnStyle}
-  background-color: #fee500;
-`;
-
-export const GoogleLogo = styled(googleLogo)`
-  width: ${pxToRem(28)};
-  height: ${pxToRem(28)};
-`;
-
-export const KakaoLogo = styled(kakaoLogo)`
-  width: ${pxToRem(28)};
-  height: ${pxToRem(28)};
-`;
-
-export const SocialLoginText = styled.span`
-  ${({ theme }) => theme.fontStyles.Body4}
-  font-weight: 700;
-  color: ${({ theme }) => theme.colors.MG_Grayscale.Gray_1};
-
-  user-select: none;
-  -webkit-user-select: none;
 `;
 
 export const SignUpSection = styled.section`
